+++ CommonLib/common.h
@@ -6,6 +6,7 @@
 #include <optional>
 #include <random>
 #include <map>
+#include <type_traits>
 
 enum class WorkingMode
 {
@@ -20,7 +21,37 @@ enum class SpeedLevel
     HIGH,
 };
 
-int getRandomInt(int low_bound, int up_bound);
+/**
+ * @brief 获取位于区间 [a, b] 内的随机整数
+ * 
+ * @tparam T 整数类型，不须填入，自动推导即可
+ * @param a 区间下界
+ * @param b 区间上界
+ * @return T 生成的随机整数
+ */
+template<typename T>
+inline T getRandomInt(T a, T b)
+{
+    static_assert(std::is_integral<T>::value, "getRandomInt 的参数类型必须是整数类型");
+    static std::random_device rd;
+    static std::mt19937 rng(rd());
+    std::uniform_int_distribution gen(a, b);
+    return gen(rng);
+}
+
+/**
+ * @brief 将枚举类型转换为对应的整形值
+ * 
+ * @tparam T 枚举类型，自动推导即可
+ * @param enum_val 需要转换的枚举类型值
+ * @return constexpr int 对应的整形值
+ */
+template <typename T>
+inline constexpr int EnumToInt(const T enum_val)
+{
+    static_assert (std::is_enum<T>::value, "EnumToInt 的参数类型必须是枚举类型");
+    return static_cast<int>(enum_val);
+}
 
 namespace Config {
 
@@ -36,14 +67,14 @@ namespace Config {
      * 
      * @return quint16 监听的端口值
      */
-    static quint16 getSlaveListenerPortValue();
+    quint16 getSlaveListenerPortValue();
 
     /**
      * @brief 设定 Listener 监听的端口
      * 
      * @param port 监听的端口
      */
-    static void setSlaveListenerPort(const quint16 port);
+    void setSlaveListenerPort(const quint16 port);
 
     /**
      * @brief 判断是否设定了 Listener 的端口
@@ -51,7 +82,7 @@ namespace Config {
      * @return true 已设定
      * @return false 未设定，需要至少调用一次 setSlaveListenerPort(port)
      */
-    static bool hasSlaveListenerPort();
+    bool hasSlaveListenerPort();
 
     /**
      * @brief 将从控机与其地址和端口进行对应
@@ -60,7 +91,7 @@ namespace Config {
      * @param host_addr 该从控机的地址
      * @param port 该从控机监听的端口
      */
-    static void addAddress(const QString &room_id, const QString &host_addr, const quint16 port);
+    void addAddress(const QString &room_id, const QString &host_addr, const quint16 port);
 
     /**
      * @brief 返回给定房间号的地址和监听端口
@@ -69,6 +100,6 @@ namespace Config {
      * @return std::pair<bool, std::pair<QString, quint16>> 若存在该房间号对应的记录，则返回 {true, {room_id, port}}, 
      * 否则返回 {false, {}}
      */
-    static std::pair<bool, std::pair<QString, quint16>> getAddress(const QString &room_id);
+    std::pair<bool, std::pair<QString, quint16>> getAddress(const QString &room_id);
 };
 #endif // COMMON_H