+++ CommonLib/commonlib.cpp
@@ -1,5 +1,0 @@
-#include "commonlib.h"
-
-CommonLib::CommonLib()
-{
-}